requirements.txt
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-classic>=0.3.0
langchain-text-splitters>=0.3.0
rank_bm25>=0.2.2
faiss-cpu>=1.9.0
pytest>=8.0.0
